1706. Where Will the Ball Fall
You have a 2-D grid of size m x n representing a box, and you have n balls.The box is open on the top and bottom sides.
Return an array answer of size n where answer[i] is the column that the ball falls out of, or -1 if the ball gets stuck in the box.