2216. Minimum Deletions to Make Array Beautiful, #Two Pointers
The array nums is beautiful if: nums.length is even. nums[i] != nums[i + 1] for all i % 2 == 0. When you delete an element,all the elements to the right of the deleted element will be shifted one unit Return the minimum number of elements to delete from nums to make it beautiful.