1284. Minimum Number of Flips to Convert Binary Matrix to Zero Matrix, #DFS
Given a m x n binary matrix mat. In one step, you can choose one cell and flip it and all the four neighbors of it if they exist (Flip is changing 1 to 0 and 0 to 1). Return the minimum number of steps required to convert mat to a zero matrix or -1 if you cannot.