2136. Earliest Possible Day of Full Bloom, #Greedy
Planting a seed takes time and so does the growth of a seed,plantTime and growTime, of length n each:
plantTime[i] is the days it takes you to plant the ith seed.Every day,you can work on planting exactly one seed.
You do not have to work on planting the same seed on consecutive days, but the planting of a seed is not complete until you have worked plantTime[i] days on planting it in total.
growTime[i] is the number of full days it takes the ith seed to grow after being completely planted.
Return the earliest possible day where all seeds are blooming.