1202. Smallest String With Swaps, #Disjoint Set, #Union Find
an array of pairs of indices where pairs[i] = [a, b] indicates 2 indices(0-indexed) of the string . You can swap the characters at any pair of indices in the given pairs any number of times.
Return the lexicographically smallest string that s can be changed to after using the swaps.