2256. Minimum Average Difference
The average difference of the index i is the absolute difference between the average of the first i + 1 elements of nums and the average of the last n - i - 1 elements.
Both averages should be rounded down to the nearest integer.
Return the index with the minimum average difference.If there are multiple such indices, return the smallest one.