2207. Maximize Number of Subsequences in a String
pattern is 2 length, eg "ab". You can add either pattern[0] or pattern[1] anywhere in text exactly once.
Note that the character can be added even at the beginning or at the end of text.
Return the maximum number of times pattern can occur as a subsequence of the modified text.