2070. Most Beautiful Item for Each Query, #Binary Search
items[i] = [pricei, beautyi] denotes the price and beauty of an item respectively.
For each queries[j], you want to determine the maximum beauty of an item whose price is less than or equal to queries[j].
If no such item exists, then the answer to this query is 0. Return an array answer of the same length as queries where answer[j] is the answer to the jth query.