916. Word Subsets
A string b is a subset of string a if every letter in b occurs in a including multiplicity.
A string a from words1 is universal if for every string b in words2, b is a subset of a.
Return an array of all the universal strings in words1. You may return the answer in any order.