563. Binary Tree Tilt, #BTree
Given the root of a binary tree, return the sum of every tree node's tilt.The tilt of a tree node is the absolute difference between the sum of all left subtree and all right subtree.If a node does not have a left child, then the sum of the left subtree node values is treated as 0. The rule is similar if the node does not have a right child.