2251. Number of Flowers in Full Bloom, #Binary Search
flowers[i] = [starti, endi] means the ith flower will be in full bloom from starti to endi (inclusive).
persons[i] is the time that the ith person will arrive to see the flowers.
Return array answer[i] is the number of flowers that are in full bloom when the ith person arrives.