1024. Video Stitching, #Greedy
clips[i] = [starti, endi] indicates that the ith clip started at starti and ended at endi.
For example, a clip [0, 7] can be cut into segments[0, 1] + [1, 3] + [3, 7].
Return the minimum number of clips needed so that we can cut the clips into segments
that cover the entire sporting event [0, time]. If the task is impossible, return -1.