1423. Maximum Points You Can Obtain from Cards, #Sliding Window
In one step, you can take one card from the beginning or from the end of the row.
You have to take exactly k cards.Your score is the sum of the points of the cards you have taken.
Given the integer array cardPoints and the integer k, return the maximum score you can obtain.