2333. Minimum Sum of Squared Difference, #Binary Search
The sum of squared difference of arrays nums1 and nums2 is defined as the sum of (nums1[i] - nums2[i])
You can modify any of the elements of nums1 by +1 or -1 at most k1 times. same for k2 on nums2;
Return the minimum sum of squared difference after modifying atmost k1 times and k2 times.